Move method to utils
